Move tiles api to own file
